Conflicts:
	WebContent/WEB-INF/views/match/managingTeam.jsp

+++ src/com/kh/santa/mypage/controller/MypageController.java
@@ -68,6 +68,9 @@ public class MypageController extends HttpServlet {
       case "mypageFollower" :
          mypageFollower(request,response);
          break;
+      case "mypagePassEdit" :
+    	  mypagePassEdit(request,response);
+           break;
       case "mypageMemberEdit" :
          mypageMemberEdit(request,response);
          break;
@@ -87,12 +90,13 @@ public class MypageController extends HttpServlet {
       case "deleteComment" :
          deleteComment(request,response);
           break;
-      case "insertFollow" :
-         insertFollow(request,response);
+      case "like" :
+    	  like(request,response);
+           break; 
+          
+      case "following" :
+         following(request,response);
           break;
-      case "deleteFollow" :
-          deleteFollow(request,response);
-           break;
       case "deleteFollowTopage" :
     	  deleteFollowTopage(request,response); 
           break;
@@ -103,7 +107,10 @@ public class MypageController extends HttpServlet {
       case "deleteMountainwish" :
     	  deleteMountainwish(request,response); 
           break;
-    	  
+          
+      case "editPass" :
+    	  editPass(request,response);
+           break;
       case "editMember" :
          editMember(request,response);
           break;
@@ -123,7 +130,8 @@ public class MypageController extends HttpServlet {
    private void mypageBoard(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
       
          Member member = (Member) request.getSession().getAttribute("authentication"); //세션조회
-         List<MemberBoard> boardList = myboardService.selectBoardDetail(member.getMemberIdx()); //조회한 세션값으로 게시판 불러오기
+         String memberIdx =member.getMemberIdx();
+         List<MemberBoard> boardList = myboardService.selectBoardDetail(memberIdx); //조회한 세션값으로 게시판 불러오기
          
          List<Object[]> res = new ArrayList<Object[]>();
         
@@ -133,14 +141,21 @@ public class MypageController extends HttpServlet {
             List<MemberBoardComment> commentList = myboardService.selectBoardComent(boardIdx); //게시판번호로 댓글찾기
             Object[] ob = new Object[] {memberBoard, file,commentList}; //객체에 담아주기
             res.add(ob);
+            
+			/*
+			 * if(!myboardService.likeBoardlist(memberIdx,boardIdx)) { //좋아요 기록이 없다면
+			 * request.setAttribute("boardlike", true); } else {
+			 * request.setAttribute("boardlike", false); }
+			 */
          }
          request.getSession().setAttribute("res", res);
+         
          request.setAttribute("mountainList", mountainList);
          wishlist = mypageService.selectMountainWishlist(member.getMemberIdx());
          request.setAttribute("wishlist", wishlist);
       request.getRequestDispatcher("/mypage/mypageBoard").forward(request, response);
    }
-   
+   //s
    //다른사람 페이지
    private void anotherBoard(HttpServletRequest request, HttpServletResponse response)throws ServletException, IOException {
           
@@ -224,6 +239,23 @@ public class MypageController extends HttpServlet {
       response.sendRedirect("/mypage/mypageBoard");
    }
    
+   //좋아요
+   private void like(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		Member member = (Member)request.getSession().getAttribute("authentication");
+		String memberIdx = member.getMemberIdx();
+		String memboardIdx = request.getParameter("membaordIdx");
+		String like = request.getParameter("boardlike");
+		System.out.println(memboardIdx);
+		System.out.println(like);
+		if(like.equals("true")) {
+			myboardService.updateBoardLike(memboardIdx);
+			request.setAttribute("boardlike", true);
+		} else {
+			myboardService.removeBoardLike(memboardIdx);
+			request.setAttribute("boardlike", false);
+		}
+		response.sendRedirect("/mypage/mypageBoard");
+	}
       
    //팔로우
    private void mypageFollow(HttpServletRequest request, HttpServletResponse response)  throws ServletException, IOException {
@@ -252,24 +284,21 @@ public class MypageController extends HttpServlet {
    
 
    
-   //팔로우 추가(다른사람 마이페이지)
-   private void insertFollow(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+   //팔로우 추가,삭제(다른사람 마이페이지)
+   private void following(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
      Member member = (Member) request.getSession().getAttribute("authentication");
+     String following= request.getParameter("follow");
      Follow follow = new Follow();
-     follow.setFollowId(request.getParameter("deletefollow"));
+     follow.setFollowId(request.getParameter("anoterIdx"));
      follow.setMemberIdx(member.getMemberIdx()); 
-     followingService.insertFollow(follow);
+     if(following.equals(following)) { //true라면 팔로우
+    	 followingService.insertFollow(follow);
+     }else{
+    	 followingService.deleteFollow(follow);
+     }
      anotherBoard(request,response);
    }
-   //팔로우 삭제(다른사람 마이페이지)
-   private void deleteFollow(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-      Member member = (Member) request.getSession().getAttribute("authentication");
-      Follow follow = new Follow();
-      follow.setMemberIdx(member.getMemberIdx());         
-      follow.setFollowId(request.getParameter("deletefollow"));   
-      followingService.deleteFollow(follow);
-      anotherBoard(request,response);
-   }
+   
    //팔로우 삭제
    private void deleteFollowTopage(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
       Member member = (Member) request.getSession().getAttribute("authentication");
@@ -280,6 +309,29 @@ public class MypageController extends HttpServlet {
       request.getRequestDispatcher("/mypage/mypageFollow").forward(request, response);
    }
    
+   //마이페이지 수정 로그인페이지
+	private void mypagePassEdit(HttpServletRequest request, HttpServletResponse response)  throws ServletException, IOException {
+	   request.setAttribute("mountainList", mountainList);
+	   request.setAttribute("wishlist", wishlist);
+	   request.getRequestDispatcher("/mypage/mypagePassEdit").forward(request, response);
+	}
+   
+
+	//마이페이지 로그인 post
+	private void editPass(HttpServletRequest request, HttpServletResponse response)throws ServletException, IOException {
+		
+		Member member = (Member) request.getSession().getAttribute("authentication");
+		String inputPassword= request.getParameter("password");
+		if(member.getUserPassword().equals(inputPassword)) { //입력한정보가 맞다면
+			request.getRequestDispatcher("/mypage/mypageMemberEdit").forward(request, response);
+		}else{
+			request.setAttribute("msg", "비밀번호가 틀렸습니다");
+			request.getRequestDispatcher("/mypage/mypagePassEdit").forward(request, response);
+		}
+		
+	}
+
+	
    //마이페이지 수정페이지
    private void mypageMemberEdit(HttpServletRequest request, HttpServletResponse response)  throws ServletException, IOException {
 	   request.setAttribute("mountainList", mountainList);
@@ -293,7 +345,7 @@ public class MypageController extends HttpServlet {
    private void editMember(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 	   
       Member member = (Member) request.getSession().getAttribute("authentication");
-      member.setNickname(request.getParameter("editnickname"));
+      String newNickname = request.getParameter("editnickname");
       member.setPhone(request.getParameter("editphone"));
       member.setEmail(request.getParameter("editmail"));
       String address = "(" + request.getParameter("zipNo") + ")" + request.getParameter("roadAddrPart1") + request.getParameter("detailaddress");
@@ -301,9 +353,9 @@ public class MypageController extends HttpServlet {
       
       if(request.getParameter("editpass")==member.getUserPassword()) { //변경함 //그대로 비번 같으면
     	  member.setUserPassword(request.getParameter("editpass"));  
-    	  mypageService.editMember(member);
+    	  mypageService.editMember(member,newNickname);
       }else {//비밀번호 변경안할때
-    	  mypageService.editMemberExclusionPassword(member);
+    	  mypageService.editMemberExclusionPassword(member,newNickname);
       }
       request.getRequestDispatcher("/mypage/mypageBoard").forward(request, response);
